+++ java-service-annotation/src/main/java/nbbrd/service/ServiceDefinition.java
@@ -26,7 +26,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Target({ElementType.TYPE})
-@Retention(RetentionPolicy.SOURCE)
+@Retention(RetentionPolicy.CLASS)
 public @interface ServiceDefinition {
 
     /**