+++ java-service-annotation/src/main/java/nbbrd/service/ServiceFilter.java
@@ -35,7 +35,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Target({ElementType.METHOD})
-@Retention(RetentionPolicy.SOURCE)
+@Retention(RetentionPolicy.CLASS)
 public @interface ServiceFilter {
 
     /**