+++ java-service-annotation/src/main/java/nbbrd/service/ServiceId.java
@@ -19,7 +19,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Target({ElementType.METHOD})
-@Retention(RetentionPolicy.SOURCE)
+@Retention(RetentionPolicy.CLASS)
 public @interface ServiceId {
 
     /**